Setup master key for account exposed endpoints on controller layer

+++ src/main/java/dev/kons/kuenyawz/controllers/AccountController.java
@@ -1,15 +1,19 @@
 package dev.kons.kuenyawz.controllers;
 
+import dev.kons.kuenyawz.configurations.ApplicationProperties;
 import dev.kons.kuenyawz.dtos.account.*;
 import dev.kons.kuenyawz.entities.Account;
+import dev.kons.kuenyawz.exceptions.UnauthorizedException;
 import dev.kons.kuenyawz.mapper.AccountMapper;
 import dev.kons.kuenyawz.services.entity.AccountService;
+import dev.kons.kuenyawz.services.logic.AuthService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.security.SecurityRequirements;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -30,6 +34,7 @@ public class AccountController extends BaseController {
 
 	private final AccountService accountService;
 	private final AccountMapper accountMapper;
+	private final ApplicationProperties properties;
 
 	@Operation(summary = "(Master) Get all accounts",
 		description = "Retrieves a list of all accounts with secure information",
@@ -41,9 +46,14 @@ public class AccountController extends BaseController {
 			schema = @Schema(implementation = ListOfAccountDto.class)
 		)
 	)
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@GetMapping
 	public ResponseEntity<Object> getAllAccounts() {
+		ensureRequesterAuthorized();
+
 		List<AccountSecureDto> accounts = accountService.getAllAccounts();
 
 		return ResponseEntity.status(HttpStatus.OK).body(new ListOfAccountDto(accounts));
@@ -69,11 +79,16 @@ public class AccountController extends BaseController {
 			)
 		)
 	)
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@PostMapping
 	public ResponseEntity<Object> createAccount(
 		@Valid @RequestBody AccountRegistrationDto accountRegistrationDto
 	) {
+		ensureRequesterAuthorized();
+
 		Account account = accountService.createAccount(accountRegistrationDto);
 		return ResponseEntity.status(HttpStatus.CREATED).body(accountMapper.fromEntity(account));
 	}
@@ -85,11 +100,16 @@ public class AccountController extends BaseController {
 			schema = @Schema(implementation = AccountSecureDto.class)
 		)
 	)
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@GetMapping("{accountId}")
 	public ResponseEntity<Object> getAccount(
 		@PathVariable Long accountId
 	) {
+		ensureRequesterAuthorized();
+
 		Account account = accountService.getAccount(accountId);
 		return ResponseEntity.ok(accountMapper.fromEntity(account));
 	}
@@ -101,23 +121,33 @@ public class AccountController extends BaseController {
 			schema = @Schema(implementation = AccountSecureDto.class)
 		)
 	)
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@PutMapping("{accountId}")
 	public ResponseEntity<Object> updateAccount(
 		@PathVariable Long accountId,
 		@Valid @RequestBody AccountPutDto accountPutDto
 	) {
+		ensureRequesterAuthorized();
+
 		Account account = accountService.updateAccount(accountId, accountPutDto);
 		return ResponseEntity.ok(accountMapper.fromEntity(account));
 	}
 
 	@Operation(summary = "Delete an account", description = "Deletes an account with the provided account ID")
 	@ApiResponse(responseCode = "204", description = "Successfully deleted account")
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@DeleteMapping("{accountId}")
 	public ResponseEntity<Object> deleteAccount(
 		@PathVariable Long accountId
 	) {
+		ensureRequesterAuthorized();
+
 		accountService.deleteAccount(accountId);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
@@ -131,37 +161,62 @@ public class AccountController extends BaseController {
 			schema = @Schema(implementation = AccountSecureDto.class)
 		)
 	)
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@PatchMapping("{accountId}/account")
 	public ResponseEntity<Object> patchAccount(
 		@PathVariable Long accountId,
 		@Valid @RequestBody AccountPatchDto accountPatchDto
 	) {
+		ensureRequesterAuthorized();
+
 		Account account = accountService.patchAccount(accountId, accountPatchDto);
 		return ResponseEntity.ok(accountMapper.fromEntity(account));
 	}
 
 	@Operation(summary = "Patch an account's password", description = "Patches with the provided request body")
 	@ApiResponse(responseCode = "204", description = "Successfully patched account's password")
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN", "USER"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@PatchMapping("{accountId}/password")
 	public ResponseEntity<Object> updatePassword(
 		@PathVariable Long accountId,
 		@Valid @RequestBody PasswordUpdateDto passwordUpdateDto
 	) {
+		ensureRequesterAuthorized();
+
 		accountService.updatePassword(accountId, passwordUpdateDto);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
 
 	@Operation(summary = "Patch an account's privilege", description = "Patches with the provided request body")
 	@ApiResponse(responseCode = "204", description = "Successfully patched account's privilege")
-	@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"})
+	@SecurityRequirements({
+		@SecurityRequirement(name = "cookieAuth", scopes = {"ADMIN"}),
+		@SecurityRequirement(name = "xApiKey")
+	})
 	@PatchMapping("{accountId}/privilege")
 	public ResponseEntity<Object> updatePrivilege(
 		@PathVariable Long accountId,
 		@Valid @RequestBody PrivilegeUpdateDto privilegeUpdateDto
 	) {
+		ensureRequesterAuthorized();
+
 		accountService.updatePrivilege(accountId, privilegeUpdateDto);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+	}
+
+	/**
+	 * Helped method to ensure the requester is authorized to perform the action.
+	 * This is specific for this controller so that direct call to Account endpoints are authenticated properly.
+	 */
+	private void ensureRequesterAuthorized() {
+		if (!(AuthService.isAuthenticatedMaster(properties) || AuthService.isAuthenticatedAdmin())) {
+			throw new UnauthorizedException("This action requires master or admin privileges");
+		}
 	}
 }
